+++ paying_for_college/config/settings/standalone.py
@@ -8,3 +8,10 @@ SECRET_KEY = "forstandaloneonly"
 STATICFILES_DIRS = (
     PROJECT_ROOT.child('local_static'),
 )
+
+HAYSTACK_CONNECTIONS = {'default':
+                        {'ENGINE':
+                         ('haystack.backends.elasticsearch_backend.'
+                          'ElasticsearchSearchEngine'),
+                         'URL': '127.0.0.1:9200',
+                         'INDEX_NAME': 'build_haystack'}}